backend/requirements.txt
pymysql
python-dotenv
pandas
fastapi
uvicorn